+++ package.json
@@ -25,11 +25,13 @@
   ],
   "dependencies": {
     "axios": "^0.16.1",
+    "md5": "^2.2.1",
     "md5.js": "^1.3.4",
     "query-string": "^4.3.4"
   },
   "devDependencies": {
     "@types/query-string": "^4.3.1",
+    "@types/md5": "^2.1.32",
     "typescript": "^2.6.1"
   }
 }
