--- yarn-error.log
@@ -1,146 +1,0 @@
-Arguments: 
-  /usr/local/Cellar/node/9.9.0/bin/node /usr/local/Cellar/yarn/1.5.1_1/libexec/bin/yarn.js add --offline md5.js
-
-PATH: 
-  /Users/skelley/Documents/Programming/shell-setup/git-number:/Users/skelley/.config/yarn/global/node_modules/.bin:/usr/local/bin:/Users/skelley/.cargo/bin:/usr/local/opt/ruby/bin:/usr/local/share/npm/bin:/Applications/Sublime Text.app/Contents/SharedSupport/bin:/usr/bin:/bin:/usr/sbin:/sbin:/opt/X11/bin:/Users/skelley/.antigen/bundles/robbyrussell/oh-my-zsh/lib:/Users/skelley/.antigen/bundles/robbyrussell/oh-my-zsh/plugins/command-not-found:/Users/skelley/.antigen/bundles/robbyrussell/oh-my-zsh/plugins/npm:/Users/skelley/.antigen/bundles/robbyrussell/oh-my-zsh/plugins/tmuxinator:/Users/skelley/.antigen/bundles/robbyrussell/oh-my-zsh/plugins/yarn:/Users/skelley/.antigen/bundles/robbyrussell/oh-my-zsh/plugins/git:/Users/skelley/.antigen/bundles/zsh-users/zsh-syntax-highlighting
-
-Yarn version: 
-  1.5.1
-
-Node version: 
-  9.9.0
-
-Platform: 
-  darwin x64
-
-npm manifest: 
-  {
-    "name": "synology-typescript-api",
-    "version": "0.3.5",
-    "description": "A Typescript implementation of a Promise-based API for the Synology DiskStation API.",
-    "main": "lib/index.js",
-    "typings": "lib/index.d.ts",
-    "scripts": {
-      "test": "tsc",
-      "prepublish": "tsc"
-    },
-    "repository": {
-      "type": "git",
-      "url": "git+https://github.com/seansfkelley/synology-typescript-api.git"
-    },
-    "license": "MIT",
-    "keywords": [
-      "synology",
-      "nas",
-      "diskstation",
-      "downloadstation",
-      "filestation",
-      "api",
-      "promise",
-      "typescript"
-    ],
-    "dependencies": {
-      "axios": "^0.16.1",
-      "md5.js": "^1.3.4",
-      "query-string": "^4.3.4"
-    },
-    "devDependencies": {
-      "@types/query-string": "^4.3.1",
-      "typescript": "^2.6.1"
-    }
-  }
-
-yarn manifest: 
-  No manifest
-
-Lockfile: 
-  # THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.
-  # yarn lockfile v1
-  
-  
-  "@types/query-string@^4.3.1":
-    version "4.3.1"
-    resolved "https://registry.yarnpkg.com/@types/query-string/-/query-string-4.3.1.tgz#59ea5c75d425817370158789aa4f6197e5abd987"
-  
-  axios@^0.16.1:
-    version "0.16.2"
-    resolved "https://registry.yarnpkg.com/axios/-/axios-0.16.2.tgz#ba4f92f17167dfbab40983785454b9ac149c3c6d"
-    dependencies:
-      follow-redirects "^1.2.3"
-      is-buffer "^1.1.5"
-  
-  debug@^2.6.9:
-    version "2.6.9"
-    resolved "https://registry.yarnpkg.com/debug/-/debug-2.6.9.tgz#5d128515df134ff327e90a4c93f4e077a536341f"
-    dependencies:
-      ms "2.0.0"
-  
-  follow-redirects@^1.2.3:
-    version "1.2.5"
-    resolved "https://registry.yarnpkg.com/follow-redirects/-/follow-redirects-1.2.5.tgz#ffd3e14cbdd5eaa72f61b6368c1f68516c2a26cc"
-    dependencies:
-      debug "^2.6.9"
-  
-  hash-base@^3.0.0:
-    version "3.0.4"
-    resolved "https://registry.yarnpkg.com/hash-base/-/hash-base-3.0.4.tgz#5fc8686847ecd73499403319a6b0a3f3f6ae4918"
-    dependencies:
-      inherits "^2.0.1"
-      safe-buffer "^5.0.1"
-  
-  inherits@^2.0.1:
-    version "2.0.3"
-    resolved "https://registry.yarnpkg.com/inherits/-/inherits-2.0.3.tgz#633c2c83e3da42a502f52466022480f4208261de"
-  
-  is-buffer@^1.1.5:
-    version "1.1.6"
-    resolved "https://registry.yarnpkg.com/is-buffer/-/is-buffer-1.1.6.tgz#efaa2ea9daa0d7ab2ea13a97b2b8ad51fefbe8be"
-  
-  md5.js@^1.3.4:
-    version "1.3.4"
-    resolved "https://registry.yarnpkg.com/md5.js/-/md5.js-1.3.4.tgz#e9bdbde94a20a5ac18b04340fc5764d5b09d901d"
-    dependencies:
-      hash-base "^3.0.0"
-      inherits "^2.0.1"
-  
-  ms@2.0.0:
-    version "2.0.0"
-    resolved "https://registry.yarnpkg.com/ms/-/ms-2.0.0.tgz#5608aeadfc00be6c2901df5f9861788de0d597c8"
-  
-  object-assign@^4.1.0:
-    version "4.1.1"
-    resolved "https://registry.yarnpkg.com/object-assign/-/object-assign-4.1.1.tgz#2109adc7965887cfc05cbbd442cac8bfbb360863"
-  
-  query-string@^4.3.4:
-    version "4.3.4"
-    resolved "https://registry.yarnpkg.com/query-string/-/query-string-4.3.4.tgz#bbb693b9ca915c232515b228b1a02b609043dbeb"
-    dependencies:
-      object-assign "^4.1.0"
-      strict-uri-encode "^1.0.0"
-  
-  safe-buffer@^5.0.1:
-    version "5.1.1"
-    resolved "https://registry.yarnpkg.com/safe-buffer/-/safe-buffer-5.1.1.tgz#893312af69b2123def71f57889001671eeb2c853"
-  
-  strict-uri-encode@^1.0.0:
-    version "1.1.0"
-    resolved "https://registry.yarnpkg.com/strict-uri-encode/-/strict-uri-encode-1.1.0.tgz#279b225df1d582b1f54e65addd4352e18faa0713"
-  
-  typescript@^2.6.1:
-    version "2.6.1"
-    resolved "https://registry.yarnpkg.com/typescript/-/typescript-2.6.1.tgz#ef39cdea27abac0b500242d6726ab90e0c846631"
-
-Trace: 
-  Error: Command failed.
-  Exit code: 1
-  Command: sh
-  Arguments: -c tsc
-  Directory: /Users/skelley/Documents/Programming/synology-typescript-api
-  Output:
-  
-      at ProcessTermError.MessageError (/usr/local/Cellar/yarn/1.5.1_1/libexec/lib/cli.js:186:110)
-      at new ProcessTermError (/usr/local/Cellar/yarn/1.5.1_1/libexec/lib/cli.js:226:113)
-      at ChildProcess.<anonymous> (/usr/local/Cellar/yarn/1.5.1_1/libexec/lib/cli.js:30281:17)
-      at ChildProcess.emit (events.js:180:13)
-      at maybeClose (internal/child_process.js:936:16)
-      at Process.ChildProcess._handle.onexit (internal/child_process.js:220:5)
